infra/datahub/requirements.txt
acryl-datahub[kafka,dbt]>=0.13
pyyaml>=6.0
